+++ packages/debug/src/common/debug-service.ts
@@ -16,7 +16,7 @@
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
-import { Disposable } from '@theia/core';
+import { Disposable, Event } from '@theia/core';
 import { ApplicationError } from '@theia/core/lib/common/application-error';
 import { IJSONSchema, IJSONSchemaSnippet } from '@theia/core/lib/common/json-schema';
 import { DebugConfiguration } from './debug-configuration';
@@ -76,6 +76,11 @@ export interface DebugService extends Disposable {
     provideDynamicDebugConfigurations?(): Promise<Record<string, DebugConfiguration[]>>;
 
     /**
+     * Provides a dynamic debug configuration matching the name and the provider debug type
+     */
+    fetchDynamicDebugConfiguration(name: string, type: string): Promise<DebugConfiguration | undefined>;
+
+    /**
      * Resolves a [debug configuration](#DebugConfiguration) by filling in missing values
      * or by adding/changing/removing attributes before variable substitution.
      * @param debugConfiguration The [debug configuration](#DebugConfiguration) to resolve.
@@ -108,6 +113,12 @@ export interface DebugService extends Disposable {
      * Stop a running session for the given session id.
      */
     terminateDebugSession(sessionId: string): Promise<void>;
+
+    /**
+     * Event handle to indicate when one or more dynamic debug configuration providers
+     * have been registered or unregistered.
+     */
+    onDidChangeDebugConfigurationProviders: Event<void>;
 }
 
 /**